refactor(domain): share virtualhost path-mapping flags
Extract the duplicated --path-to-app / --path-to-url / --path-to-container
flag definitions and the IngressPath[] building logic from the
`virtualhost create` and `virtualhost update` commands into a shared
module at src/lib/resources/domain/virtualhost/flags.ts.
